詳細表示の行作成をしている2つのクラスを一つに統合
